3.0-agent-under-the-hood/pyproject.toml
[project]
name = "3-0-agent-under-the-hood"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.18",
    "langchain-openai>=1.2.1",
    "python-dotenv>=1.2.2",
]
